+++ skfp/fingerprints/estate.py
@@ -35,7 +35,7 @@ class EStateFingerprint(FingerprintTransformer):
 
         X = np.array([FingerprintMol(x) for x in X])
         if self.variant == "bit":
-            X = (X[:, 0] > 0).astype(int)
+            X = X[:, 0] > 0
         elif self.variant == "count":
             X = (X[:, 0]).astype(int)
         else:  # "sum" variant
